+++ R/spiro_import.R
@@ -586,7 +586,7 @@ spiro_anonymize <- function(info) {
   }
 
   # replace personal information by anonymized id
-  id <- get_id(
+  id <- get_anonid(
     name = info$name,
     surname = info$surname,
     birthday = birthday
@@ -603,7 +603,7 @@ spiro_anonymize <- function(info) {
 
 #' Get the anonymization id from personal data
 #'
-#' \code{get_id()} returns the anonymization id corresponding to given personal
+#' \code{get_anonid()} returns the anonymization id corresponding to given personal
 #' data.
 #'
 #' By default, the spiro package anonymizes personal information obtained from
@@ -611,10 +611,10 @@ spiro_anonymize <- function(info) {
 #' The default anonymization ensures that no personal information is
 #' accidentally revealed, e.g. by sharing spiro outputs as .Rda files.
 #'
-#' While there is no way to directly deanonymize the data, get_id() allows you
-#' to recreate the ids, when meta data (name, surname and birthday) are known.
-#' Birthday is only used within the id generation if available in the original
-#' raw data.
+#' While there is no way to directly deanonymize the data, get_anonid() allows
+#' you to recreate the ids, when meta data (name, surname and birthday) are
+#' known. Birthday is only used within the id generation if available in the
+#' original raw data.
 #'
 #' To disable the anonymization process during import use spiro(anonymize =
 #' FALSE)
@@ -630,8 +630,8 @@ spiro_anonymize <- function(info) {
 #' @return A character string, containing the anonymized id.
 #'
 #' @examples
-#' get_id("Jesse", "Owens", "12.09.1913")
+#' get_anonid("Jesse", "Owens", "12.09.1913")
 #' @export
-get_id <- function(name, surname, birthday = NULL) {
+get_anonid <- function(name, surname, birthday = NULL) {
   digest::digest(c(name, surname, birthday), "crc32")
 }